Sun, Sep 10, 2023, 12:21 PM -03:00

+++ cs50/aula4/filter/helpers.c
@@ -1,6 +1,6 @@
 #include "helpers.h"
 
-void copying(BYTE avrg, RGBTRIPLE *rgbt);
+void copying(DWORD avrg, RGBTRIPLE *rgbt);
 
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
@@ -15,7 +15,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for(int w = 0; w < width; w++)
         {
-            BYTE avrg = (image[h][w].rgbtBlue + image[h][w].rgbtGreen + image[h][w].rgbtRed) / 3;
+            DWORD avrg = (image[h][w].rgbtBlue + image[h][w].rgbtGreen + image[h][w].rgbtRed) / 3; // it will colapse 255 byte of 8 bits. need a long byte, if it exists.
             copying(avrg, &image[h][w]);
             // copy(&avrg, &image[h][w].rgbtGreen);
             // copy(&avrg, &image[h][w].rgbtRed);
@@ -52,7 +52,7 @@ void edges(int height, int width, RGBTRIPLE image[height][width])
     return;
 }
 
-void copying(BYTE avrg, RGBTRIPLE *rgbt)
+void copying(DWORD avrg, RGBTRIPLE *rgbt)
 {
     // *rgbt = *avrg;
     *rgbt.rgbtBlue = avrg;
